Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float i;
-    scanf("%f",&i);
-    printf("You entered: %.4f",i);
+    double i;
+    scanf("%lf",&i);
+    printf("You entered: %.4lf",i);
     return 0;
 }
